Refactor commands into modular subpackages
Split the monolithic commands.py into dedicated submodules for completion, providers, and template functionality. Added src/promptheus/_provider_data.py for shared provider utilities. Introduced a completions subpackage for shell completion logic and templates. Updated setup.py to a minimal version deferring configuration to pyproject.toml. This improves maintainability and modularity of the codebase.
